scripts/terraform/cloudfunctions/ce/src/python/requirements.txt
# Function dependencies, for example:
# package>=version
google.cloud
google.cloud.bigquery
google.cloud.storage
google.cloud.scheduler
google-cloud-secret-manager
google-cloud-bigquery-datatransfer
boto3==1.17.24
google-cloud-pubsub==2.5.0
google-api-python-client
bigquery-schema-generator==1.5